Refactor new detail in all categories
