+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/detail/categories/creatures/CreaturesItemDetailContainer.composable.kt
@@ -1,5 +1,5 @@
 @file:OptIn(ExperimentalFoundationApi::class)
-package com.example.zeldacompendium.presentation.ui.detail.categories.equipment
+package com.example.zeldacompendium.presentation.ui.detail.categories.creatures
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.PaddingValues
@@ -10,15 +10,17 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.example.zeldacompendium.data.remote.responses.ItemDetailModel
+import com.example.zeldacompendium.presentation.ui.commons.HorizontalPages.pages.CookingEffectPage
+import com.example.zeldacompendium.presentation.ui.commons.HorizontalPages.pages.DropsPage
+import com.example.zeldacompendium.presentation.ui.commons.HorizontalPages.pages.HealingPage
 import com.example.zeldacompendium.presentation.ui.commons.HorizontalPages.pages.LocationPage
-import com.example.zeldacompendium.presentation.ui.commons.HorizontalPages.pages.PropertiesPage
 
 @Composable
-fun EquipmentItemDetailContainer(
+fun CreaturesItemDetailContainer(
    itemInfo: ItemDetailModel,
    gameId: Int
 ) {
-   val pagerState = rememberPagerState(pageCount = { 2 })
+   val pagerState = rememberPagerState(pageCount = { 3 })
    HorizontalPager(
       modifier = Modifier.fillMaxHeight(),
       state = pagerState,
@@ -28,9 +30,17 @@ fun EquipmentItemDetailContainer(
          vertical = 12.dp
       ),
    ) { page ->
-         when (page) {
-            0 -> PropertiesPage(itemInfo, gameId)
-            1 -> LocationPage(itemInfo, gameId)
+         if (!itemInfo.data.edible) {
+            when (page) {
+               0 -> DropsPage(itemInfo)
+               1 -> LocationPage(itemInfo, gameId)
+            }
+         } else {
+            when (page) {
+               0 -> HealingPage(itemInfo)
+               1 -> CookingEffectPage(itemInfo)
+               2 -> LocationPage(itemInfo, gameId)
+            }
          }
-   }
+      }
 }
